# optimize NullableColumn::serialize_batch for FixedLengthColumnBase