--- a/be/src/column/column.h
+++ b/be/src/column/column.h
@@ -211,8 +211,26 @@ public:
         return 0;
     };
 
-    virtual void serialize_batch_with_null_vector(uint8_t* dst, Buffer<uint32_t>& slice_sizes, size_t chunk_size,
-                                                  uint32_t max_one_row_size, uint8_t* null_vector) {}
+    virtual void serialize_batch_with_null_masks(uint8_t* dst, Buffer<uint32_t>& slice_sizes, size_t chunk_size,
+                                                 uint32_t max_one_row_size, uint8_t* null_masks, bool has_null) {
+        uint32_t* sizes = slice_sizes.data();
+
+        if (!has_null) {
+            for (size_t i = 0; i < chunk_size; ++i) {
+                memcpy(dst + i * max_one_row_size + sizes[i], &has_null, sizeof(bool));
+                sizes[i] += 1 + serialize(i, dst + i * max_one_row_size + sizes[i]);
+            }
+        } else {
+            for (size_t i = 0; i < chunk_size; ++i) {
+                memcpy(dst + i * max_one_row_size + sizes[i], null_masks + i, sizeof(bool));
+                sizes[i] += 1;
+
+                if (null_masks[i]) {
+                    sizes[i] += serialize(i, dst + i * max_one_row_size + sizes[i]);
+                }
+            }
+        }
+    }
 
     // deserialize one data and append to this column
     virtual const uint8_t* deserialize_and_append(const uint8_t* pos) = 0;