--- a/be/src/column/nullable_column.cpp
+++ b/be/src/column/nullable_column.cpp
@@ -213,9 +213,8 @@ size_t NullableColumn::serialize_batch_at_interval(uint8_t* dst, size_t byte_off
 
 void NullableColumn::serialize_batch(uint8_t* dst, Buffer<uint32_t>& slice_sizes, size_t chunk_size,
                                      uint32_t max_one_row_size) {
-    for (size_t i = 0; i < chunk_size; ++i) {
-        slice_sizes[i] += serialize(i, dst + i * max_one_row_size + slice_sizes[i]);
-    }
+    _data_column->serialize_batch_with_null_vector(dst, slice_sizes, chunk_size, max_one_row_size,
+                                                   _null_column->get_data().data());
 }
 
 const uint8_t* NullableColumn::deserialize_and_append(const uint8_t* pos) {