Add `serialize_batch_with_null_vector` to base class `Column`, and implement an optimized version for `FixedLengthColumnBase`.

Close #619.

----
The information of benchmark test is as follows:
- Dataset: [New York Taxi Data](https://clickhouse.com/docs/en/getting-started/example-datasets/nyc-taxi/).
- The number of rows: 12_9897_9494.
- Query SQL:
``` SQL
-- Q3
SELECT passenger_count, year(pickup_date) AS year_, count(*) 
FROM trips_mergetree 
GROUP BY passenger_count, year_;

-- Q4
SELECT passenger_count, year(pickup_date) AS year_, round(trip_distance) AS distance, count(*) 
FROM trips_mergetree 
GROUP BY passenger_count, year_, distance 
ORDER BY year_, count(*) DESC;
```

`passenger_count` is `int8`, `trip_distance` is `float64`, and the  return type of `year()` is nullable `int32`.



The followings are the cost time before and after the optimization:

Q3:
| Parallel | Before (ms) | After (ms) |
| -------- | ----------- | ---------- |
| 1        | 6989        | 6271       |
| 2        | 3558        | 3253       |
| 4        | 1913        | 1676       |
| 8        | 1321        | 1256       |
| 16       | 1037        | 936        |

Q4:
| Parallel | Before (ms) | After (ms) |
| -------- | ----------- | ---------- |
| 1        | 7860        | 8017       |
| 2        | 4143        | 4159       |
| 4        | 2148        | 2191       |
| 8        | 1673        | 1623       |
| 16       | 1417        | 1381       |

And the cost time rate of `NullableColumn<T>::serialize_batch` before and after the optimization are 9.36% and 5.42%, respectively.